refactor: 불필요한 코드 제거, 이름 수정
